Move TestTerminal to TestCore project
