+++ src/AppMotor.CliApp/Terminals/ITerminal.cs
@@ -8,6 +8,10 @@ namespace AppMotor.CliApp.Terminals;
 /// no interaction with the terminal window - like setting the terminal's title, cursor position or obtaining
 /// the terminal's width. For advanced features, see <see cref="ITerminalWindow"/>.
 /// </summary>
-public interface ITerminal : ITerminalOutput, ITerminalInput
+/// <remarks>
+/// This interface primarily exists for unit testing purposes where no actual terminal/console may be available.
+/// See "TestTerminal" in the "AppMotor.TestCore" package for an implementation.
+/// </remarks>
+public interface ITerminal : ITerminalInput, ITerminalOutput
 {
 }
