+++ src/AppMotor.TestCore/AppMotor.TestCore.csproj
@@ -23,6 +23,7 @@
   </ItemGroup>
 
   <ItemGroup>
+    <ProjectReference Include="..\AppMotor.CliApp\AppMotor.CliApp.csproj" />
     <ProjectReference Include="..\AppMotor.Core\AppMotor.Core.csproj" />
   </ItemGroup>
